+++ packages/@runlightyear/airtable/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/airtable
 
+## 1.2.12
+
+### Patch Changes
+
+- Updated dependencies [494972892]
+  - @runlightyear/lightyear@1.17.0
+
 ## 1.2.11
 
 ### Patch Changes