+++ packages/@runlightyear/gmail/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/gmail
 
+## 0.6.13
+
+### Patch Changes
+
+- Updated dependencies [494972892]
+  - @runlightyear/lightyear@1.17.0
+
 ## 0.6.12
 
 ### Patch Changes