+++ packages/@runlightyear/gsheets/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/gsheets
 
+## 0.8.12
+
+### Patch Changes
+
+- Updated dependencies [494972892]
+  - @runlightyear/lightyear@1.17.0
+
 ## 0.8.11
 
 ### Patch Changes