+++ packages/@runlightyear/linear/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/linear
 
+## 0.10.12
+
+### Patch Changes
+
+- Updated dependencies [494972892]
+  - @runlightyear/lightyear@1.17.0
+
 ## 0.10.11
 
 ### Patch Changes