+++ packages/@runlightyear/postmark/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/postmark
 
+## 0.5.12
+
+### Patch Changes
+
+- Updated dependencies [494972892]
+  - @runlightyear/lightyear@1.17.0
+
 ## 0.5.11
 
 ### Patch Changes