+++ packages/@runlightyear/smtp/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/smtp
 
+## 0.3.12
+
+### Patch Changes
+
+- Updated dependencies [494972892]
+  - @runlightyear/lightyear@1.17.0
+
 ## 0.3.11
 
 ### Patch Changes