+++ packages/@runlightyear/zoom/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/zoom
 
+## 0.9.13
+
+### Patch Changes
+
+- Updated dependencies [494972892]
+  - @runlightyear/lightyear@1.17.0
+
 ## 0.9.12
 
 ### Patch Changes